Code review 2.
