+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/basicLogicalOperators.scala
@@ -2421,57 +2421,3 @@ object AsOfJoin {
   }
 }
 
-object NearestByJoin {
-  /** Upper bound on `numResults`. Mirrors the K-overload limit of `MaxMinByK`. */
-  val MaxNumResults: Int = 100000
-}
-
-/**
- * A logical plan for a nearest-by top-K ranking join. For each row on the left side it returns
- * up to `numResults` rows from the right side ordered by `rankingExpression`:
- *   - `NearestByDistance`: smallest values of `rankingExpression` first.
- *   - `NearestBySimilarity`: largest values of `rankingExpression` first.
- *
- * The `approx` field records the user's APPROX/EXACT choice from the SPIP. Today both modes
- * use the same brute-force rewrite. The flag is preserved on the logical plan so future
- * indexed approximate-nearest-neighbor strategies can fire only when `approx = true`,
- * leaving EXACT queries unaffected. See the SPIP linked from SPARK-56395.
- */
-case class NearestByJoin(
-    left: LogicalPlan,
-    right: LogicalPlan,
-    joinType: JoinType,
-    approx: Boolean,
-    numResults: Int,
-    rankingExpression: Expression,
-    direction: NearestByDirection)
-  extends BinaryNode with SupportsNonDeterministicExpression {
-
-  require(Seq(Inner, LeftOuter).contains(joinType),
-    s"Unsupported nearest-by join type $joinType")
-
-  // APPROX permits a nondeterministic ranking expression (per the SPIP); the rewrite
-  override def allowNonDeterministicExpression: Boolean = approx
-
-  // Right-side attributes are always declared nullable because the rewrite materializes them
-  // through `Inline` over `MaxMinByK`'s `ArrayType(.., containsNull = true)`, which widens
-  // every struct field to nullable. Declaring them nullable here keeps the analyzed schema
-  // consistent with the optimized plan (and with what users see in cached or written outputs).
-  override def output: Seq[Attribute] =
-    left.output ++ right.output.map(_.withNullability(true))
-
-  def duplicateResolved: Boolean = left.outputSet.intersect(right.outputSet).isEmpty
-
-  override lazy val resolved: Boolean = {
-    childrenResolved &&
-      expressions.forall(_.resolved) &&
-      duplicateResolved
-  }
-
-  final override val nodePatterns: Seq[TreePattern] = Seq(NEAREST_BY_JOIN)
-
-  override protected def withNewChildrenInternal(
-      newLeft: LogicalPlan, newRight: LogicalPlan): NearestByJoin = {
-    copy(left = newLeft, right = newRight)
-  }
-}
